+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,9 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+
+0.110.0
+=======
 - ui: stop storing "title" attributes into BasketEntryModel objects, also remove when touching objects
 - middleware: add tool "browser_database_tool" for manipulating browser database dumps (json)
 - middleware: don’t list email addresses for newsletter if user is tagged with “newsletter:opt-out”